raspi3--22-oled-display/requirements.txt
Adafruit-Blinka==5.2.3
adafruit-circuitpython-busdevice==5.0.1
adafruit-circuitpython-framebuf==1.3.3
adafruit-circuitpython-ssd1306==2.8.3
Adafruit-GPIO==1.0.3
Adafruit-PlatformDetect==2.14.3
Adafruit-PureIO==1.1.5
board==1.0
Pillow==7.2.0
pyftdi==0.51.2
pyserial==3.4
pyusb==1.0.2
rpi-ws281x==4.2.4
RPi.GPIO==0.7.0
spidev==3.5
sysv-ipc==1.0.1
